Code refactor for the application.

+++ src/main/java/pl/wasik/damian/spring/app/fleet/management/system/fleet/models/VehicleHire.java
@@ -1,5 +1,6 @@
 package pl.wasik.damian.spring.app.fleet.management.system.fleet.models;
 
+
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,6 +9,7 @@ import org.springframework.format.annotation.DateTimeFormat;
 import pl.wasik.damian.spring.app.fleet.management.system.parameters.models.Client;
 import pl.wasik.damian.spring.app.fleet.management.system.parameters.models.Location;
 
+
 import java.util.Date;
 
 @Entity
@@ -15,38 +17,38 @@ import java.util.Date;
 @NoArgsConstructor
 @AllArgsConstructor
 public class VehicleHire {
-
-    @Id
+		
+	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "vehicleid", insertable = false, updatable = false)
-    private Vehicle vehicle;
-    private Integer vehicleid;
-
-    @DateTimeFormat(pattern = "yyyy-MM-dd")
-    private Date dateOut;
-
-    private String timeOut;
-
-    @DateTimeFormat(pattern = "yyyy-MM-dd")
-    private Date dateIn;
-
-    private String timeIn;
-
-    @ManyToOne
-    @JoinColumn(name = "clientid", insertable = false, updatable = false)
-    private Client client;
-    private Long clientid;
-
-    @ManyToOne
-    @JoinColumn(name = "locationid", insertable = false, updatable = false)
-    private Location location;
-    private Long locationid;
-
-    private String price;
-
-    private String remarks;
+	@Column(name="id")
+	private int id;
+	
+	@ManyToOne
+	@JoinColumn(name="vehicleid", insertable=false, updatable=false)
+	private Vehicle vehicle;
+	private Integer vehicleid;
+	
+	@DateTimeFormat(pattern = "yyyy-MM-dd")
+	private Date dateOut;
+	
+	private String timeOut;
+	
+	@DateTimeFormat(pattern = "yyyy-MM-dd")
+	private Date dateIn;
+	
+	private String timeIn;
+	
+	@ManyToOne
+	@JoinColumn(name="clientid", insertable=false, updatable=false)
+	private Client client;
+	private Integer clientid;
+	
+	@ManyToOne
+	@JoinColumn(name="locationid", insertable=false, updatable=false)
+	private Location location;
+	private Integer locationid;
+	
+	private String price;
+	
+	private String remarks;
 }
