+++ src/main/java/pl/wasik/damian/spring/app/fleet/management/system/fleet/repositories/VehicleTypeRepository.java
@@ -1,9 +1,11 @@
 package pl.wasik.damian.spring.app.fleet.management.system.fleet.repositories;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import pl.wasik.damian.spring.app.fleet.management.system.fleet.models.VehicleType;
 
 @Repository
-public interface VehicleTypeRepository extends JpaRepository<VehicleType, Long> {
+public interface VehicleTypeRepository extends JpaRepository<VehicleType, Integer> {
+
 }
